Migrated Asset values from Double to BigDecimal

+++ app/src/main/java/com/ceaver/assin/assets/Asset.kt
@@ -1,9 +1,11 @@
 package com.ceaver.assin.assets
 
+import java.math.BigDecimal
+
 data class Asset(//
         val name: String,
         val symbol: String,
-        val amount: Double,
-        val btcValue: Double,
-        val usdValue: Double
+        val amount: BigDecimal,
+        val btcValue: BigDecimal,
+        val usdValue: BigDecimal
 )
